+++ packages/identity-frontend/src/components/pages/Home.page.tsx
@@ -6,7 +6,11 @@ import {
   shortcutsData,
   noscriptData
 } from "../../data";
-import { ActionBar, OverviewCard, Paper } from "@libresat/frontend-components";
+import {
+  ActionBar,
+  ExpandablePaper,
+  Paper
+} from "@libresat/frontend-components";
 import { IPage } from "../../types";
 import { NavLink } from "react-router-dom";
 import { Shell, IShellProps } from "@libresat/frontend-components";
@@ -32,7 +36,7 @@ const HomePage = (props: IPage) => (
         ...homeData.actionBar.search
       }}
     />
-    <OverviewCard {...homeData.overview}>
+    <ExpandablePaper {...homeData.overview}>
       <Paper data-grid={{ x: 0, y: 0, w: 7, h: 3 }}>
         <p>3 Scopes</p>
       </Paper>
@@ -42,7 +46,18 @@ const HomePage = (props: IPage) => (
       <Paper data-grid={{ x: 0, y: 4, w: 12, h: 3 }}>
         <p>2 Roles</p>
       </Paper>
-    </OverviewCard>
+    </ExpandablePaper>
+    <ExpandablePaper {...homeData.relationships}>
+      <Paper data-grid={{ x: 0, y: 0, w: 12, h: 3 }}>
+        <p>scope-1 -> user-1 -> role-1</p>
+      </Paper>
+      <Paper data-grid={{ x: 0, y: 4, w: 12, h: 3 }}>
+        <p>scope-2 -> user-2 -> role-2</p>
+      </Paper>
+      <Paper data-grid={{ x: 0, y: 8, w: 12, h: 3 }}>
+        <p>scope-3 -> user-3 -> role-3</p>
+      </Paper>
+    </ExpandablePaper>
   </Shell>
 );